Assertion Changes

+++ src/main/java/com/rpg/service/impl/GameActionServiceImpl.java
@@ -63,7 +63,7 @@ public class GameActionServiceImpl {
 			} else {
 				System.out.println("You're fighting a: Troll");
 			}
-			Integer earnedExperience = (int) (rand.nextInt() * 10 + 1);
+			Integer earnedExperience =rand.nextInt() * 10 + 1;
 			rpgChar.setExperience(earnedExperience);
 			rpgCharacterService.updateCharacter(rpgChar);
 			System.out.println("Your Character has gained experience");
